syntax overhaul: sort out move semantics for Header & Image
Involved removing a lot of the const specifiers around the code.
Hopefully this won't prevent the compiler from performing certain
optimisations - to be benchmarked. Also removes some safety checks that
were in place to prevent unintended modifications to Header prior to
calling get_image(), but this risks of doing this should be obvious to
programmers.

With these changes, using Headers and Images should be easier in a range
of different contexts, for instance using them directly in STL
containers.