GUI::MRView::Image: Split class in half
GUI::MRView::Image now derives from ImageBase, which provides some (but not all) of the functionality previously stored in the Image class.
This split will allow deriving a different class from ImageBase, that contains a scratch buffer rather than a buffer for an image on disk.
This in turn will allow overlays for images that are computed dynamically.