MRView: Restore ODF preview as separate dialog window
The ODF preview GL window will now spawn (when desired) as a separate window.
This way, the image overlay is promoted as the primary functionality, but the single ODF preview is still available; furthermore, by being a separate window, it does not affect the sizing of the ODF overlay toolbar.
It also means you can have a very large preview if you wish, eihter on a second monitor or just by switching between it and the main window.
Related to #219.