Implemented Marching Cubes algorithm
The Marching Cubes algorithm generates a triangulated mesh representation of a three-dimensional image.
It provides smoother mesh representations than the previous implementation of the vox2mesh() function, which intentionally meshed the faces between adjacent voxels where the input mask image transitioned from a 1 to a 0.
The mask2mesh command has also been renamed to mrmesh, and a couple of command-line options added, to highlight how this approach can be applied to non-binary images.