Connectome tool: Modify UI tabbing
The node and edge visualisation labels at the left edge of the tool now occupy 2 grid columns.
This allows additional controls, which only appear as new rows when certain parameters are selected, to have an indentation of half the width of the main labels, giving them adequate width but separating them from the primary controls.