Connectome tool: Visualisation warnings
Due to the design of the connectome toolbar, it is possible that some users may become confused when they attempt to change how the connectome elements are drawn; this is particularly true for edges, which default to not being drawn at all.
This change introduces warning icons that pop up if the uesr tries to change some fundamental property of connectome display, but no change will be observed because the visibility of the relevant element is set to 'none'.