Connectome tool: Enhanced transparency
If a polygonal object is drawn with transparency enabled, draw the object twice, but with the face culling winding convention reversed the first time.
This will draw all away-facing polygons before all front-facing polygons, making the far side of the object visible through transparency but without requiring an explicit distance sort, and should provide an appropriate polygon rendering order in most cases.
Also allow a reduced specular component to appear on the far side, based on the level of transparency enabled.