Connectome tool: Make better use of connectome2tck
Initially, the streamline / streamtube rendering in the connectome tool was achieved by running tcknodeextract to get a track file per edge, opening the directory path in the connectome tool, and the exemplar paths would be calculated on-the-spot by finding the track files in that directory.
Now, the user must first run connectome2tck with the -exemplars option and -files single; this will create a single track file containing the exemplar path for every edge in the connectome. This file is then opened in the connectome tool.