OpenGL apps: fix issues with ODF preview tool on MacOSX
Turns out this was due to calling makeCurrent() manually, probably
getting in the way of Qt's own handling of context switching. I'm
guessing this causes the front/back buffer swap to happen on the wrong
window, since that gets called by Qt outside of the paintGL() function,
and hence after we'd set the main window as current... Turns out Qt
handles all of this just fine as-is, so I've just removed those
overrides for the painGL(), resizeGL() & initializeGL() functions.