ODF preview tool: instantiate the preview tool on demand
This moves the instantiation of the preview tool outside of the ODF
tool constructor, and into its show_preview_slot().

In general, I try to limit the amount of instantiation of unnecessary
objects as much as possible, and delay their creation until they're
explicitly needed. The idea is to make sure MRView starts quickly, to
make sure it's fast to use from the command-line, etc. This isn't such a
big deal since the ODF tool is not shown by default, so the constructor
doesn't get invoked until the user requests it, but hey...