Better handling of streamline lengths
Previously, the length of each streamline was quantified as the product of the trqactography step size, and the number of track segments (one less than the number of points).
However, one situation where this is no longer accurate (and is prominent now that it is the default behaviour) is when down-sampling is used during tractography.
As the two streamline endpoints are guaranteed to be included in the down-sampled streamline, the two segments at either end of the track may be shorter than the other segments, resulting in an over-estimation of length using the old heuristic.
This commit introduces functions to calculate the streamline length in the presence of such an effect.

The update affects the following:
- tckedit: Application of length constraints.
- tckstats: Statistics now calculated using exact lengths rather than from the histogram; streamlines will be assigned to histogram based on calculated length rather than number of points.
- tck2connectome: Length-based edge metrics will now use length in mm rather than a step size count.
- tckmap: length and invlength contrasts.