Connectome tool: Better handlling of element visibility
Previously, a single boolean was used to toggle whether a particular element (node or edge) was visible.
Now, in addition to this, both the physical size and the transparency of each element is also tested.
In addition to not calling the draw functions for these elements, it also means that if either nodes are drawn based on degree, or edges are drawn based on the relevant nodes being visible, then the visibility of the former will react dynamically as the size / transparency of the latter is modified.