Render cubes using redundant vertices
Previously, an efficient scheme for drawing cubes was used, whereby the provoking vertex framework was used to provide normals for faces, removing the need to store redundant vertex locations and face normals.
However in the context of the connectome tool, this required annoying code branching in order to disable interpolation of normals.
Instead, now each vertex and normal is simply duplicated, such that each face appears flat despite interpolation of the normals.