Connectome tool: Modify nodes using input matrix files
Previously, if some visual property were to be manipulated using raw data, this was only possible using a vector file.
Now, the user can import a matrix file, and select their node of interest; the relevant row of the imported matrix will then be used to manipulate the relevant visual property.
For now, this selector is just a spinbox for each setting; later a single node list view will likely be used in place of these controls.