syntax overhaul: sort out save() & display() functions
These are no longer members of the image class, allowing them to be used
for any image-like object (image or adapter, etc.). In addition, if the
image is accessed using direct IO, the save() method will use the new
Image::dump_to_mrtrix_file() function, which bypasses copying into a
temporary RAM buffer prior to write-out - this should obviate the need
for the BufferScratchDump class used in tckmap.

This also introduces functions to test whether objects are images,
headers, or adapters, for use with SFINAE. These are:
- enable_if_image_type<ImageType>()
- enable_if_header_type<ImageType>()
- is_header_type<ImageType>()
- is_image_type<ImageType>()
- is_pure_image<ImageType>()
- is_adapter_type<ImageType>()

See the implementation of the save() function for example usage.