Connectome tool: Support multi-node selection and node property by matrix
In situations where a matrix file is used to control node visibility / colour / size / transparency, the manual selection of a node determines the row of the matrix to be sampled in order to set the relevant property of each node.
These changes add compatibility for this capability with selection of multiple nodes; comboboxes allow the user to set the statistic (minimum, mean, sum, maximum) used to combine the values recorded from multiple rows of the matrix to produce a scalar result.