Colour maps: Provide non-GLSL mapping
In addition to GLSL code as a string, the basic colour maps now also provide a function to convert an amplitude between 0.0 and 1.0 to a colour based on that map.
This is used in particular in the connectome tool, where commonly multiple facets are drawn for each connectome element and therefore pre-computing these colours is preferable.
It also allows the node list view in the connectome tool to update the colour previews when such maps are used.