GUI apps: fix issues with progressbar with Qt >= 5.4
This is likely to be a recurring problem. Basically, Qt 5.4's
OpenGLWidget now shares its context with the rest of the GUI. What this
means is that updating the progressbar (i.e. a different window
altogether, with a different context) during a screen update changes the
current context, and subsequent operations in the screen update happen
in the wrong context. This tries to catch that situation and handle it
appropriately.

Note that there is also the issue that updating the progressbar
inherently causes Qt to process all events, which I think causes it to
issue a new update for the main OpenGL window, which is itself already
in the middle of an update. This causes this message to appear:

    QWidget::repaint: Recursive repaint detected

It doesn't seem to harm actual operation, but we may need to refactor
things to avoid this type of problem.