build: temporary files now stored in separate folder
This changes the location of temporary files (objects, executables,
library, etc) from within the source tree (potentially named after the
corresponding config if not default), to a dedicated folder per config.
The config file for each config now also resides in that folder, so that
the folder name is the config name. Unless a named config is requested
in ./configure, by default this folder will be called 'default'.  This
now means that executables will now reside in default/bin/ rather than
bin/.

The reason for these changes is simplify management of co-existing
configs - each one now resides in its own folder, making it trivial to
clean it out if needed. It also makes it easier to request building of
specific target, since the config can now easily be parsed from the
folder location, e.g.:

    $ ./build debug/bin/mrconvert