(Failed) attempt at more efficient looping construct
Committed for reference. It is more efficient than the current Loop
class, but is probably still not fully adequate since it relies on the
body of the loop being provided using a lambda function - and g++ won't
always inline a lambda function (and you can't force it to...). This
results in a measurable performance difference in for example tckgen.