Connectome tool: Improvements to node selection
When performing manual selection of nodes, the connectome tool now distinguishes between nodes / edges that are 'associated' with the selected nodes (ie. an edge that is connected to one selected node, or a node that has a visible edge connection to a selected node), as opposed to nodes / edges that are in no way involved with the selection. Full settings for controlling how these elements appear visually are also provided.