MRView: prevent repaint while painting
As discussed in #293, #295, #296 & #298. This essentially ensures the
ProgressBar display only kicks in after 1s (so will only affect things
for large loads); that the main window does not try to process further
updates in the meantime; that mouse events to the main window are not
processed until completion; and that the GL context is preserved only if
one in found.

This works as expected on Qt4. On Qt5, for some reason the background of
the main window is cleared to default (grey) as soon as the progressbar
is displayed. I'm hoping this is a glitch in Qt5.5 - I can't see any
reason for them to clear the background of a widget that explicitly has
updates disabled...