Vector plot: Change thresholding
Thresholds were being applied to the fixel size when this parameter was being used to scale the length of the drawn lines, and the associated value when line lengths were either unity or scaled by the associated value.
This change makes these thresholds apply to the fixel associated value in all circumstances.
In addition, application of these thresholds was moved from the fragment shader to the geometry shader, which should be faster.