ACT: Modify back-tracking algorithm
Previously, whenever back-tracking was enabled, a simple integer counter was incremented and the streamline was shortened by that length.
This issue with this approach was that if back-tracking was enabled early during propagation, but successful in finding a new streamline trajectory, another back-tracking event further down the trajectory would immediately have a large revert step.
This modification essentially resets the back-tracking length to 1 step whenever a new back-tracking event is encountered, giving the algorithm a better chance of successfully traversing multiple difficult regions.