tckgen: Two null-distribution algorithms
'Nulldist1' now behaves the same as the old 'Nulldist' algorithm, i.e. Euler integration with random directions.
The new 'Nulldist2' algorithm performs tracking in a similar manner to iFOD2 (i.e. along arcs), but using non-informative FODs; it should therefore be used to capture the null distribution when the iFOD2 tracking algorithm is to be used.
Closes #150.