QSpinBox: Prevent unwanted double-changes
The presence of long processing times being triggered by changing the value in a QSpinBox would sometimes cause the value to be changed twice (i.e. double-increment or double-decrement).
This occurred due to the QSpinBox allowing the user to hold the mouse button down to quickly cycle through values; but if the processing for each change took too long, the timeout event to trigger the cycling behaviour would appear before the mouse release event in the event queue.
This change simply processes all other events when the timeout occurs, then re-checks the mouse state to see if the user is in fact still holding down the mouse button and therefore wants to cycle through values.
Closes #306.