Connectome tool: Implement connectome list
A QListView is now used to store a set of connectome matrices of interest.
The node and edge visual properties can be set to 'Connectome', and this will use the currently-selected connectome to control that visual property; this is distinct from the 'Matrix file' option, which sets the relevant visual property according to the data in some file but this does not change when a different connectome within the list is selected.
Another benefit of this approach is that the 'edge visibility' and 'edge colour' default to 'Connectome', so once the user has initialised the tool using the parcellation image ad loaded a connectome, they immediately have a useful visualisation that can then be further customizes.
Closes #320.