different approach for thread-local RNG
This is to get around the lack of support for C++11's thread_local on
MacOSX. This instead tries to use the less standard __thread keyword to
allocate a thread-local pointer, which has to be set to point to the
thread's actual RNG while the thread is running. The actual RNG's are
allocated per-thread as members of the thread functor class.