sparse/image.h: Check explicit inclusion of lib/image.h
Although this file includes the line "#include "image.h", the compiler incorrectly identifies this as the file itself rather than loading lib/image.h. Consequently, the necessary information on class MR::Image is absent, and the programmer gets a set of not-entirely-useful errors.
This will hopefully give the programmer a hint as to how to fix the issues.