DWI::FMLS::Segmenter: Change handling of merged lobes
By default, the FMLS segmenter will always split the FOD into multiple lobes when there are multiple peaks.
The 'ratio_of_peak_value_to_merge' parameter allows this to be changed, by merging near-indistinguishable peaks to be merged into a single lobe - it' based on the ratio of the maximal value in the trough between the peaks, to the minimal peak amplitude of the peaks in question.
Previously this merge operation was postponed until the entire FOD had been traversed; however this was found to cause unwanted behaviour.
This merging of peaks is now dealt with as the FMLS process progresses.
'Retrospective assignments' (directions that are to be assigned to a particular lobe, but are not done so until the end of the FMLS process; this is to prevent larger lobes from 'invading' smaller ones, instead providing a clean bisection between them) are updated accordingly as the merges take place.