mrconvert: reset default scaling parameters for floating-point output
For non-floating-point outputs, the scaling parameters default to those
of the input header, to preserve the information content. As discussed
in #346, these parameters are not required for floating-point images,
and if anything prevent subsequent commands from optimising their
processing, since non-default scaling forces an explicit pre-load to RAM
(if requested by the application), that would otherwise have been
avoided using memory-mapping.

The worry with this commit was that converting a strongly scaled integer
image to floating-point, and then back to integer would result in
information loss due to discretization since the scaling parameters
present in the original header would have been lost when converting to
floating-point. However, this is likely to be a very rare issue, given
that these conversions only happen by explicit request from the user
(using the -datatype option).