MRView: new ASSERT_GL_MRVIEW_CONTEXT_IS_CURRENT macro
This checks that the main MRView glarea context is current (only when
using Qt >= 5.4). This has been added to verify that the context is as
expected just about everywhere I could think of. Hopefully, this'll
allow a debug (or at least -assert) build to catch any issues related to
using the wrong GL context.

There is also an associated ASSERT_GL_CONTEXT_IS_CURRENT(widget) macro
to check that the widget's context is current, which can be used when
other contexts than the main MRView window are current (e.g. the ODF
preview tool).