Connectome tool: Fix clamping button ranges
For the node controls, if e.g. the value set for the maximum was less than that for the minimum, it would be clamped to the current value of the minimum; but the valueChanged() signal emitted by the corresponding clamping would cause an infinite recursion.
For the edge controls, this clamping was not yet implemented.