tckmap and tckdynamicmap: -zero_outside_fov option
By default, when the TWI track-wise statistic is based on the streamline endpoints, and a streamline endpoint is outside the image FoV, the last point along the streamline that is within the image FoV will be used.
This option allows for the TWI factor of such streamlines to instead be set to zero.