DWI::Directions::Set: Better direction adjacency detection
Previously, determining that two directions on the unit sphere were 'adjacent' involved a simple heuristic: Generating the midpoint between the two directions, then performing an exhaustive search looking for a direction in the set closer to that midpoint than the two directions in question.
This new implementation uses the QuickHull algorithm to obtain the convex hull of the direction set, and defines direction adjacency based on the resulting triangulation.
This should not have a profound influence on current apps that require this definition (e.g. the FOD FMLS segmenter), but is required for the MRview Dixel-ODF visualisation.