Scripts: Add citation details
Much like the binary commands, the scripts will now print the relevant citation details at the end of the help page.
This has been implemented in such a way that new scripts and new articles can be easily added.
If a script contains relevant citations, and the script is run in normal or verbose mode, a note will be printed to the user at the commencement of the script reminding them of the fact.
The scripts will now also print the full help page if there is an error in parsing the arguments and options, rather than the default brief error message.