MRview ODF tool: Speedup
Previously, the flat surface shading of dixel-based ODFs was achieved by duplicating vertices, i.e. for each triangle all three vertices were written explicitly, and the normal was computed on the CPU and duplicated for all three vertices.
This commit instead uses indexing-based triangulation of the data (as is done for SH-based ODFs), and surface normals are calculated using a geometry shader.
This provides some speedup of rendering by reducing the amount of data transfer to the GPU.