Standardize ProgressBar behaviour
Previously, wherever a progress bar was created, the programmer may or may not have added an ellipsis to the end of the message, and there may or may not have been a space after the ellipsis (which offsets the cursor).
This commit removes this variability, and therefore makes the overall MRtrix terminal appearance more consistent, by handling the ellipsis display within the ProgressInfo class.
If a functor is used to update the text, however, no ellipsis is displayed.
Additionally, the ellipsis disappears once the relevant processing is completed.
Closes #30.