Scripts: Convenience function for finding eddy binary
I noticed in my FSL 5.0.9 download that there's no eddy binary; instead there's eddy_cuda and eddy_openmp.
This would make the distortion correction scripts fail.
A new convenience function, getFSLEddyPath(), checks for the appropriate command name.
In addition to this, if the user requests it, the CUDA version of eddy will be used.
Furthermore, getPEAxis() has been changed to getPEDir(), and will store both the axis number, and whether or not the sign of the PE axis is reversed with respect to MRtrix's RAS convention; this is then fed through to the topup / eddy configuration file.
Closes #406.