tckmap/tckdynamicmap: Backtracking
Previously, if a streamline endpoint was outside the fMRI image, by default the first point inside the FoV would be selected.
Now, by default a value of 0 will be assigned to the streamline; however, if the -backtrack option is provided, then it's not just the first point inside the FoV, but the first with a valid time-series.
In addition, streamlines should no longer get a value of NaN in cases where -backtrack is not provided.