tcksample: -use_tdi_fraction
In this mode, tcksample assumes that the intensity in each voxel is an additive quantity, that is to be distrbuted between the streamlines traversing it (based on the length of each streamline through that voxel).
This is done by first generating a TDI, and then multiplying the value contributed to the streamline in each voxel by (streamline_length_in_voxel / track_density_in_voxel).