5ttgen: -sgm_amyg_hipp option
Option to represent the amygdalae and hippocampi as sub-cortical grey matter in the 5TT image.
In the freesurfer algorithm, this just changes the tissue labels of the relevant structures.
In the FSL algorithm, these structures will be segmented using FIRST and inserted into the 5TT image just as is done with other sub-cortical grey matter structures.