Changes to connectome LUT / config handling
Handling of connectome node information is no longer done using a combination of lookup tables and 'config files' (which specified the desired index for each node of interest).
Instead, these data are handled entirely using lookup tables.
The command labelconvert (formerly labelconfig) will read an input LUT (on which the input parcellated image is based), and an output LUT (on which the output parcellated image should be based), and figure out the index permutations from those.
This interface is simplified by the fact that the format of LUT files will now be determined automatically.
This also simplifies the mrview connectome tool, as the user must only select the LUT corresponding to the parcellation image they have used, without worrying about the presence or absence of a config file responsible for node index permutation.
Related to #373.