labelsgmfix: Update for compatibility with connectome_no_config branch
In the prior implementation of this script, it would happily read from either the FreeSurfer lookup table file, or an MRtrix connectome config file, since the first two columns were identical (index then name).
With the config file concept removed, this was no longer possible; the script needs to be compatible with lookup tables in multiple formats (e.g. the default FreeSurfer LUT provided with MRtrix adds an abbreviated node name in the second column.
Therefore, this script must now exploit the labelconvert command in order to determine the appropriate indices for the structures being replaced.