mrcalc: Fix 'operation not permitted' error
This arose from the same image being input multiple times to the one operation, which would make use of the list of already-opened images. However, in this case, StackEntry::image_is_complex would remain uninitialized, which could by chance be non-zero; this in turn would make the output image complex, which could cause subsequent operations to fail if not compatible with complex data.
Closes #518.