mrview ODF tool: Fix preview in dixel mode
Amplitude data sent to the preview renderer was not being filtered based on the selected slice, resulting in noisy ODF shapes as the wrong amplitude would be read for each direction.
Also enable / disable the level of detail control in the preview window according to whether the data is dixel-based (and hence whether this control has any meaning).
Closes #500.
Also modified so that b=0 shells no longer appear in the list of available shells to display.