File::MMap: use memory-mapping for RW images if held on local FS
This detects the filesystem that the file is held on, and uses
memory-mapping for RW files if the FS isn't reported as NFS, CIFS or
FUSE (which should catch SSHFS & GlusterFS). Otherwise the current
delayed write-back approach is used. This also checks whether the FS is
mounted with the synchronous flag, and swiches to delayed write-back in
this case too, since that would also be expected to kill performance.