bvecs: fix import/export
As discussed in #506. While the bvecs format is relative to the image axes,
it also assumes that the images are stored using a LHS coordinate
system, as per the old Analyze format. In other words, for an image
stored in NIfTI-standard RAS, the bvecs are interpreted as if the image
was stored in LAS. Another way to put it (although much more ambiguous)
is that when interpreting bvecs, the image should be assumed
radiological even if it is actually stored neurological.