MRview: Windowing reset fixes
When the user requests an intensity windowing reset, this will now be based on the min/max of the current slice if locked to axes, or the min/max of the volume if not.
In addition, if the current intensity windowing is invalid, the first non-empty slice encountered will force a windowing reset, as discussed in: http://community.mrtrix.org/t/trouble-with-nibabel-generated-nii-file/102/12.
Closes #405.