tcksample: Remove integral statistic
When the -precise option is used with the mean track-wise statistic, the units of the per-streamline factors is (quantitative measure) per mm; therefore, this can be multiplied by the track lengths obtained with tckstats -dump to obtain the integral per streamline.
Furthermore, it may enable similar estimates to be derived using statistics other than the mean.