dwipreproc: Check header transform of topup output
In the original script, the header of the input image was imported into the field map output image of topup; this is because topup was erroneously writing that image with an identity transform.
Now, the transform of this image will be explicitly checked; and if it's not identical to that of the input image to topup, the user will be presented with a warning, and the script will perform the transform replacement.