Scripts: New library function getOutputPath
This removes the need to use os.path.join(lib.app.workingDir, ...) whenever an output file is written.
Additionally, a boolean flag is provided to the function to control whether or not quotation marks should be placed around the path; this is required in runCommand() calls to properly handle user-defined paths that include whitespace.