Scripts: Rename getOutputPath() to getUserPath()
This function should be used whenever the user provides a path to a file. It will handle inserting the location of the working directory, as well as placing quotation marks around the path if required (when the path contains a space, and when the path is being used in a runCommand() call).