mrview tractography tool: Fix error with colouring by endpoints
Loading a scalar file for tractogram colouring from some other directory would change the working directory, which in turn would mean that the loaded track file could not be located for the purpose of calculating tangents between streamline endpoints.
This is mitigated by instead calculating these tangents when the track file is first read, and then producing the GPU buffers when this colouring mechanism is first requested.