SIFT model: Disable heuristic GM scaling by default
In the SIFT model, it is common for the FODs in voxels containing part WM and part GM to be under-represented by streamlines, as the streamlines are only permitted to exist within some fraction of the voxel volume.
A heuristic was provided to compensate for this by using the ACT 5TT image to down-scale the FOD size in such voxels.
However with the emergence of more tissue-specific spherical deconvolution approaches, this heuristic is no longer required in all circumstances.
Therefore, this heuristic scaling can now be activated via a command-line option, rather than the previous behaviour where it could be de-activated via an option.
Closes #608.