Avoid 'using namespace X' at namespace scope in headers
This includes all members of namespace X in any file that includes
that header and imports the enclosing namespace itself. For example:

In header.h:

    namespace DWI {
      using namespace Eigen
      ...

In myapp.cpp:

    using namespace DWI;
    Stride::set (...);    <-- causes conflict with Eigen::Stride namespace

It's OK to do this in cpp files (since they don't get included
elsewhere), but still probably good practice not to - it keeps the code
clear as to which function we're talking about. It's also OK to use it
within function scope, since it won't leak outside of that scope.