Surface::Freesurfer::read_label() function
This function reads in a FreeSurfer .label file, resulting in a vector of vertices (which will typically be a subset of all vertices in the original mesh), and a Surface::Scalar (one quantitative value associated with each vertex).
Also changed Surface::Scalar to derive from Eigen::Array rather than having it as a member.