Simplified handling for IO rediction
As discussed in #686. All IO redirection is now handled the same whether
in append mode or not. Progress updates that don't modify the text (i.e.
don't invoke ProgressBar::update()) will show a progress meter similar
to [=========], with each character corresponding to 2%. No progress is
shown for a simple busy indicator. For updates that do modify the text,
a full line is shown for each update, but the rate of update is
gradually reduced from its normal full rate by doubling the interval
between updates each time - i.e. only updates 0,1,2,4,8,16,... are
shown.