FOD FMLS segmenter: Store multiple peaks per lobe
In the case where the FMLS 'peak ratio to merge' parameter is less than 1.0, peaks that do not have an adequately deep trough between them will be merged into a single lobe.
This change allows the 'FOD_lobe' class to track the orientations of multiple FOD peaks when this occurs.
This will not affect default behaviour in any way, since 1.0 is the default value for this parameter; but it may be useful in other applications making use of this method.
Calculation of appropriate weights for numerical integration of FODs has also been moved to a discrete class in order to enable external access.