New command: mrhistmatch
This command determines a non-linear mapping in order to match the histogram of one image to another image.
One particular use case of interest is taking the reciprocal of the b=0 image, then matching the histogram to a T1 image, to produce a pseudo-T1 DWI that can be used for registration to T1 using the sum-of-squares metric.
This commit also includes a substantial amount of work in homogenizing the generation and use of histograms by different commands; previously some functionality was duplicated in different areas, sometimes with minor differences.
Code branching design of mrstats has changed; hopefully this will make it more amenable to improvements in the future.
Methods that generate a histogram will now by default use the Freedman-Diaconis rule for determining an appropriate granularity for the histogram.
Closes #671.