DICOM: Initial extraction of additional useful information
Extraction of fields such as phase encode direction / sign, echo time, and effective echo spacing is preliminary, and should not be trusted until it has been tested on a wider array of data. Conversion of this data into the quantity required by FSL eddy is also currently ambiguous and should therefore not be attempted.