dwipreproc: Initial commit of redesign
Rather than configuring topup and eddy solely based on user input, the new philosophy of this script is to capture the phase encoding information from the image header(s) (which should have been captured from the initial DICOM import), and call topup / eddy appropriately.
The previous functionality of the dwipreproc script is retained (with some small changes to the interface) by providing options that enable the user to manually specify the acquisition protocol design, direction and readout time of EPI phase encoding, in cases where this information is not available in the image header.
Additionally, if such options are used but the phase encoding information IS available in the header, the user will be warned if their manual settings differ from those present in the header.
Note that this is very much an initial commit of this script, and has not been tested all all, let alone comprehensively.