dwipreproc: First attempt at volume recombination
Since the script now aims to operate on any arbitrary organisation of image data, it must use the explicit determination of matching volume pairs, along with their phase encoding information, to derive the corresponding jacobians and perform volume recombination.
Note that this update is UNTESTED and unlikely to work currently.