fixelconvert: New option -template
When converting from old to new format, enable nomination of an existing fixel image in the new format that is to be used as a template for performing the conversion.
If the number and directions of fixels in each voxel match, then this will ensure that the output values are created with the appropriate offsets such that they can be used in conjunction with the template fixel image.