memory alignment for Eigen: proposed solution
This introduces a framework for dealing with over-aligned types that
Eigen 3.3 provides, which are necessary to benefit from modern CPUs' AVX
instructions. The issue applies to any class that might be dynamically
allocated (via new/delete) and contains a static Eigen class that
requires over-alignment (currently Vector4d, Matrix4f, Matrix2d,
Matrix4d, transform_type, as far as I can tell). If dynamically
allocated, alignment is not guaranteed (this is an issue that goes to
the heart of the C++ standard), but this cannot be checked at
compile-time. This makes it almost impossible to provide any guarantees
that our code is safe since the testing framework only checks a limited
range of the possibilities in the code, leading to random runtime
failures.

Eigen can verify alignment of the relevant classes within their
constructors, but that will only catch issues at runtime. All this buys
us is an immediate abort, and a more interpretable error message when
this happens. But still no compile-time guarantees...

So this commit attempts to provide a solution to this. It requires ALL
classes to include the new MEMALIGN() macro, which will force the use of
a over-aligned operator new() method, but crucially this will only force
alignment for classes that genuinely need it (as reported by alignof(X)
> default_alignment). For classes that don't need alignment, the default
::operator new() function is called, and hopefully any overhead
optimised out (these should all be compile-time decisions).

See cmd/test_alignment.cpp for a demonstration of this in use. Note that
this command will be removed before merging to master, this is just a
means of documenting the logic and testing done in the history for later
inspection if required.

The final piece of the puzzle will be to provide a means of inspecting
the code to detect classes or structs that do not have MEMALIGN()
specified. A script to do this will be added in a later commit.